Let Format_# functions take output of functions that return string
VAR_IN_OUT CONSTANT can only take string literals or variables and not
the return of a function. Switching to VAR_INPUT will take more ram in
some cases, but will make the general case much easier to handle.

We also fleshed out the rest of the test cases for the Format_#
functions. :^)